Update README.md
Project Goal: The goal of this project is to build a movie recommendation system that can recommend movies to users based on their past ratings and preferences.
Data: The data for this project will be a dataset of movie ratings and reviews. This data can be collected from a variety of sources, such as IMDb, Rotten Tomatoes, and Netflix.
Algorithms: There are two main types of algorithms that can be used to build a movie recommendation system: content-based filtering and collaborative filtering. Content-based filtering recommends movies that are similar to movies that the user has already rated highly. Collaborative filtering recommends movies that other users who have similar ratings to the user have also rated highly.
Implementation: The movie recommendation system will be implemented in Python. The Python library scikit-learn can be used to implement both content-based filtering and collaborative filtering algorithms.
Evaluation: The movie recommendation system will be evaluated by its ability to recommend movies that the user has not rated before. This can be done by asking the user to rate a set of movies that the system has recommended.